chore(gw): refactor gateway auth tests into the new testing framework
